LPS-53266 Wiki fails because a JSP references WikiUtil.getFormattedContent() and it can't be found
